requirements/main.txt
aiofiles==24.1.0
aiohttp==3.11.15
aiosqlite==0.21.0
alembic==1.15.2
anyio==4.9.0
bcrypt==4.3.0
fakeredis==2.28.0
fastapi-pagination==0.12.34
fastapi[all]==0.115.12
faststream[redis,cli]==0.5.37
hiredis==3.1.0
orjson==3.10.16
psycopg[binary]==3.2.6
puremagic==1.28
pyJWT[crypto]==2.10.1
pydantic-settings==2.8.1
pydantic>=2.11.1
python-multipart==0.0.20
python-ulid==3.0.0
python_dotenv==1.1.0
redis==5.2.1
slowapi
taskiq-dependencies==1.5.7
taskiq-redis==1.0.3
taskiq==0.11.16
tenacity<10.0.0
typer<1
uvicorn[standard]==0.34.0
uvloop==0.21.0; sys_platform != 'win32'
virtualenv==20.29.3
waldiez==0.4.2
zipstream==1.1.4
